+++ mpaas-20201028/alibabacloud_mpaas20201028/models/_create_open_single_data_response_body.py
@@ -7,11 +7,15 @@ from darabonba.model import DaraModel
 class CreateOpenSingleDataResponseBody(DaraModel):
     def __init__(
         self,
+        access_denied_detail: str = None,
+        http_status_code: int = None,
         request_id: str = None,
         result: bool = None,
         result_code: str = None,
         result_message: str = None,
     ):
+        self.access_denied_detail = access_denied_detail
+        self.http_status_code = http_status_code
         self.request_id = request_id
         self.result = result
         self.result_code = result_code
@@ -25,6 +29,12 @@ class CreateOpenSingleDataResponseBody(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.access_denied_detail is not None:
+            result['AccessDeniedDetail'] = self.access_denied_detail
+
+        if self.http_status_code is not None:
+            result['HttpStatusCode'] = self.http_status_code
+
         if self.request_id is not None:
             result['RequestId'] = self.request_id
 
@@ -41,6 +51,12 @@ class CreateOpenSingleDataResponseBody(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AccessDeniedDetail') is not None:
+            self.access_denied_detail = m.get('AccessDeniedDetail')
+
+        if m.get('HttpStatusCode') is not None:
+            self.http_status_code = m.get('HttpStatusCode')
+
         if m.get('RequestId') is not None:
             self.request_id = m.get('RequestId')